requirements.txt
beautifulsoup4==4.12.0
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.15
Werkzeug==2.2.3
